fix: add correcao formulario e add verificacao delete

+++ src/usuario/delete_usuario.php
@@ -10,18 +10,33 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
         exit;
     }
 
-    $sql = "DELETE FROM usuario WHERE id_usuario = :id_usuario";
-
-    $stmt = $pdo->prepare($sql);
-    $stmt->bindParam(':id_usuario', $id_usuario);
-
     try {
+        $sqlVerifica = "select id_usuario from aluno where id_usuario = :id_usuario";
+        $stmtVerifica = $pdo->prepare($sqlVerifica);
+        $stmtVerifica->bindParam(':id_usuario', $id_usuario);
+        $stmtVerifica->execute();
+        
+        $existeAluno = $stmtVerifica->fetch(PDO::FETCH_ASSOC);
+        
+        if($existeAluno){
+            echo "<script>
+                alert('ERRO: Não é possível deletar este usuario pois ele está vinculado a um aluno!');
+                window.location.href = '../pages/dados_usuario.php';
+            </script>";
+            exit;
+        }
+
+        $sql = "DELETE FROM usuario WHERE id_usuario = :id_usuario";
+        $stmt = $pdo->prepare($sql);
+        $stmt->bindParam(':id_usuario', $id_usuario);
+
         if ($stmt->execute()) {
-            header('Location: ../pages/dados_usuario.php');
+            header('Location: ../pages/dados_usuario.php?sucesso=deletado');
             exit;
         } else {
             echo 'Erro ao deletar o usuario';
         }
+        
     } catch (PDOException $e) {
         echo "Erro: " . $e->getMessage();
     }
